% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{anomalies_analysis}
\alias{anomalies_analysis}
\title{Analisys of anomalies in data}
\usage{
anomalies_analysis(sensorData, windowLength, numSamples, numCores)
}
\arguments{
\item{sensorData}{A list of numbers containing a time series}

\item{windowLength}{Size of the samples to be generated}

\item{numSamples}{Number of samples to extract from the original dataset}

\item{numCores}{Number of cores to use to run parallel code}
}
\value{
List of elements that contain information about the outcomes of the calculations done
}
\description{
This function analyses a dataset by executing the algorithms for variation analysis and outliers
detection. Several samples are extracted from the dataset and analysed by using sliding windows, so it is
possible to detect anomalies as the samples are varying.
}
\examples{
sensorData <- rnorm(250, mean=15)
anomalies_analysis(sensorData, 20, 3, 2)

}
