% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{exp_diff}
\alias{exp_diff}
\title{Exponential difference transformation}
\usage{
exp_diff(inputData)
}
\arguments{
\item{inputData}{A list of numbers containing a time series}
}
\value{
List containing the input data transformed with the exponential difference transformation
}
\description{
This function takes a time series as an input and it transforms the data in such a way it
highlights the extreme values of the time series. It is based on the diff transformation, applying an exponential
factor according to the formula: yi(x) = (xi-x(i-1)) * 1.1^(abs(xi-x(i-1)) - 15)
}
\examples{
myList <- sample (1:5,10,replace=TRUE)
exp_diff(myList)
}
