% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{outliers_analysis}
\alias{outliers_analysis}
\title{Outliers analysis}
\usage{
outliers_analysis(sensorData)
}
\arguments{
\item{sensorData}{A list of numbers containing a time series}
}
\value{
List of elements that contain information about the outcomes of the calculations done.
}
\description{
Function that perform several analyses to check whether data from a sensor seems to contain outliers.
It calculates the Grubbs test for individual outliers. Additionally, it calculates several homogeneity tests:
SNHT, Pettitt, Buishand R, Buishand U and Lanzante. It makes use of the transformations provided with this
package, as well as the outliers detector for calculating angles based on the outcomes generated by the
homogeneity tests' statistics.
}
\examples{
sensorData <- rnorm(150, mean=15)
outliers_analysis(sensorData)

}
