% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{poly_reg}
\alias{poly_reg}
\title{Polynomial regression transformation}
\usage{
poly_reg(inputData)
}
\arguments{
\item{inputData}{A list of numbers containing a time series}
}
\value{
List containing the input data transformed with the polynomial regression transformation
}
\description{
This function takes a time series as input and it transforms the data using polynomial regression.
First of all, it creates a prediction model with polynomial regression of grade 5 and, then, it generates a
prediction based on the model. Finally, it subtracts the prediction from the original time series.
}
\examples{
myList <- sample (1:5,10,replace=TRUE)
poly_reg(myList)

}
