% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectors.R
\name{turning_points_analysis}
\alias{turning_points_analysis}
\title{Angles calculation in turning points}
\usage{
turning_points_analysis(testData)
}
\arguments{
\item{testData}{A list of numbers containing a time series. Expected to be the result of a homogeneity test}
}
\value{
List of elements that contain information about the number of angles, their position and calculation
}
\description{
Angles calculation in turning points
}
\examples{
sensorData <- sample (1:10,30,replace=TRUE)
turning_points_analysis(sensorData)
}
