% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{variation_analysis}
\alias{variation_analysis}
\title{Variation analysis}
\usage{
variation_analysis(sensorData)
}
\arguments{
\item{sensorData}{A list of numbers containing a time series}
}
\value{
List of elements that contain information about the outcomes of the calculations done.
}
\description{
Function that calculates all the variation tests for checking whether data from a sensor seems to be
generated randomly, or if it contains too much noise. It calculates Coefficient of Variation, IQR, Runs test and
Ljung-Box statistical test.
}
\examples{
sensorData <- sample (1:10,30,replace=TRUE)
variation_analysis(sensorData)

}
